requirements.txt
aiohttp>=3.8.0
pydantic>=2.0.0
python-dotenv>=0.19.0
pydantic-ai>=0.0.1
click>=8.0.0